% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create.views}
\alias{create.views}
\title{Create Views}
\usage{
create.views(views = c("traitsview", "yieldsview", "mgmtview"))
}
\arguments{
\item{views}{names of views to create (must have assocaites 'view'.sql in inst/extdata}

\item{con}{database connection}
}
\value{
nothing, creates views in database as a side effect
}
\description{
create planting, seeding, coppice views
}
\details{
Loads and submits scripts to generate views. *views.sql files are located in inst/extdata and can be accessed with \code{system.file("traitsview.sql", package = "db")}
}
\examples{
\dontrun{create.views(views = c("mgmtview"))}
}
\author{
David LeBauer
}
\references{
Stack Overflow question \href{http://stackoverflow.com/questions/11418920/how-to-send-a-query-from-a-sql-file/}{how to send a query from a .sql file?}
}

