% Generated by roxygen2 (4.0.2): do not edit by hand
\name{db_copy}
\alias{db_copy}
\title{db_copy}
\usage{
db_copy()
}
\value{
nothing, copies ebi_production db to ebi_analysis db on ebi-forecast server
}
\description{
Backup ebi_production to ebi_analysis
}
\details{
Executes the db_copy.sh script on ebi-forecast
}
\author{
David LeBauer
}

