% Generated by roxygen2 (4.0.2): do not edit by hand
\name{expect_covs}
\alias{expect_covs}
\title{count traits with no covariates}
\usage{
expect_covs(var, cov)
}
\arguments{
\item{var}{}

\item{cov}{}
}
\value{
result of expect_equal
}
\description{
Check that traits have required covariates
}
\author{
David LeBauer
}

