% Generated by roxygen2 (4.0.2): do not edit by hand
\name{jagify.dt}
\alias{jagify.dt}
\title{JagifyDT}
\usage{
jagify.dt(dt)
}
\arguments{
\item{dt}{input trait data}
}
\value{
dt as data.table result transformed to meet requirements of PEcAn meta-analysis model
}
\description{
Prepare trait data for JAGS meta-analysis
}
\details{
Convert queried data to format required by JAGS meta-analysis model
}
\author{
David LeBauer
}

