% Generated by roxygen2 (4.0.2): do not edit by hand
\name{newdate}
\alias{newdate}
\title{newdate}
\usage{
newdate(date)
}
\arguments{
\item{date}{}
}
\value{
transformed date
}
\description{
Remove 0000's, 00's from dates
}
\details{
replaces unknown years -> NA; unknown month -> Jan; unknown day -> 01
}
\author{
David LeBauer
}

