% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pull.data.tables}
\alias{pull.data.tables}
\title{Pull Data Tables}
\usage{
pull.data.tables(...)
}
\arguments{
\item{...}{arguments passed to \code{\link{pull.tables}}}
}
\value{
list of BETYdb tables imported as data.tables
}
\description{
Import tables from BETYdb as data.tables
}
\author{
David LeBauer
}

