% Generated by roxygen2 (4.0.2): do not edit by hand
\name{query.traitdata}
\alias{query.traitdata}
\title{Query data by trait}
\usage{
query.traitdata(trait)

query.traitdata(trait)
}
\arguments{
\item{trait}{}

\item{trait}{}

\item{con}{}
}
\value{
data frame with species, mean, se, and n

data frame with species, mean, se, and n
}
\description{
Query all data from traitsview by trait

.. content for \descrption{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
David LeBauer

David LeBauer
}

