% Generated by roxygen2 (4.0.2): do not edit by hand
\name{query.yield.data}
\alias{query.yield.data}
\title{Extract yield data from BETYdb}
\usage{
query.yield.data(spstr)
}
\arguments{
\item{spstr}{is the species.id integer or string of}
}
\value{
dataframe ready for use in meta-analysis
}
\description{
query.yield.data extracts yield data from BETYdb for a given set of species,
converts all statistics to summary statistics, and prepares a dataframe for use in meta-analysis.
}

