% Generated by roxygen2 (4.0.2): do not edit by hand
\name{query.yields}
\alias{query.yields}
\title{Query Yields}
\usage{
query.yields(spstr)
}
\arguments{
\item{spstr}{}

\item{...}{}
}
\value{
queries yield data for a given species string
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
David LeBauer
}

