% Generated by roxygen2 (4.0.2): do not edit by hand
\name{transform.nas.dt}
\alias{transform.nas.dt}
\title{transform NA's}
\usage{
\method{transform}{nas.dt}(dt)
}
\arguments{
\item{dt}{input data}
}
\value{
dt with NA's replaced
}
\description{
Function to remove NA values from database queries
}
\author{
David LeBauer
}

